move limiter logic into pkg
